:recycle: Move .NET contracts into equally named subfolder
